Implement retrieval of the fields name in different cases
