+++ src/Lightweight/SqlQuery/Select.hpp
@@ -249,21 +249,21 @@ inline LIGHTWEIGHT_FORCE_INLINE SqlSelectQueryBuilder& SqlSelectQueryBuilder::Fi
     if constexpr (sizeof...(MoreRecords) == 0)
     {
         Reflection::EnumerateMembers<FirstRecord>(
-            [&]<size_t FieldIndex, typename FieldType>() { Field(Reflection::MemberNameOf<FieldIndex, FirstRecord>); });
+            [&]<size_t FieldIndex, typename FieldType>() { Field(FieldNameAt<FieldIndex, FirstRecord>); });
     }
     else
     {
         Reflection::EnumerateMembers<FirstRecord>([&]<size_t FieldIndex, typename FieldType>() {
             Field(SqlQualifiedTableColumnName {
                 .tableName = RecordTableName<FirstRecord>,
-                .columnName = Reflection::MemberNameOf<FieldIndex, FirstRecord>,
+                .columnName = FieldNameAt<FieldIndex, FirstRecord>,
             });
         });
 
         (Reflection::EnumerateMembers<MoreRecords>([&]<size_t FieldIndex, typename FieldType>() {
              Field(SqlQualifiedTableColumnName {
                  .tableName = RecordTableName<MoreRecords>,
-                 .columnName = Reflection::MemberNameOf<FieldIndex, MoreRecords>,
+                 .columnName = FieldNameAt<FieldIndex, MoreRecords>,
              });
          }),
          ...);
